refactor: caminho adicionar psicologo ajustado
