+++ src/pages/dashboard/psicologos/adicionarPsicologo/adicionarPsicologo.css
@@ -7,14 +7,19 @@
     align-items: center;
     justify-content: space-around;
     height: 100vh;
-    width: 60%;
+    width: 70%;
     border: 3px solid var(--DarkestBlueFy);
     border-radius: 24px;
     margin: 0 auto;
+    background-color: var(--LightGrayFy);
 
     h2 {
         font-size: 2rem;
         margin-bottom: 20px;
         font-weight: bold;
+    }
+
+    form{
+        padding: 0em 2em;
     }
 }
